Restructuring, added sources folder andd moved LC Clients
