+++ sunpy/net/unifieddownloader/__init__.py
@@ -1,9 +1,10 @@
-from .factbase import UnifiedDownloader
-from .eve import EVEClient
-from .lyra import LYRAClient
-from .norh import NoRHClient
-from ..vso import VSOClient
-from .RHESSI import RHESSIClient
+from sunpy.net.unifieddownloader.downloader_factory import UnifiedDownloader
+from sunpy.net.unifieddownloader.sources.eve import EVEClient
+from sunpy.net.unifieddownloader.sources.lyra import LYRAClient
+from sunpy.net.unifieddownloader.sources.norh import NoRHClient
+from sunpy.net.vso.vso import VSOClient
+from sunpy.net.unifieddownloader.sources.rhessi import RHESSIClient
+from sunpy.net.unifieddownloader.sources.noaa import NOAAIndicesClient,NOAAPredictClient
 UnifiedDownloader.register(EVEClient,EVEClient._can_handle_query)
 UnifiedDownloader.register(LYRAClient,LYRAClient._can_handle_query)
 UnifiedDownloader.register(NoRHClient,NoRHClient._can_handle_query)
